+++ Tools/com.renoise.AutoMate.xrnx/changelog.md
@@ -2,7 +2,8 @@
 
 ## 0.55
 
-- Add integration with keybindings (passive mode)
+- Add Cut action
+- Add keybindings for all basic actions (Cut,Copy,Paste,Clear), in all scopes
 
 ## 0.5